backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
supabase==2.13.0
python-dotenv==1.1.0
